# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ME Piston Component Running Hours Record API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vessels/{vesselId}/monthly-rh:
    get:
      operationId: listMonthlyRh
      tags: [monthly-rh]
      summary: List monthly RH log entries (last 120)
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly RH log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRhEntry"
    post:
      operationId: addMonthlyRh
      tags: [monthly-rh]
      summary: Add a monthly RH log entry
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonthlyRhInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRhEntry"
        "400":
          description: Validation error
  /vessels/{vesselId}/monthly-rh/{entryId}:
    patch:
      operationId: updateMonthlyRhEntry
      tags: [monthly-rh]
      summary: Update a monthly RH log entry
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonthlyRhInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRhEntry"
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deleteMonthlyRhEntry
      tags: [monthly-rh]
      summary: Delete a monthly RH entry
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Components ────────────────────────────────────────────────────────────
...
  /vessels/{vesselId}/cylinders:
    get:
      operationId: listCylinders
      tags: [cylinders]
      summary: List all cylinders for a vessel
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cylinder list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cylinder"
  /vessels/{vesselId}/cylinders/{cylinderNumber}:
    patch:
      operationId: updateCylinder
      tags: [cylinders]
      summary: Update a cylinder's assignment and overhaul records
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: cylinderNumber
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CylinderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cylinder"

  # ─── Movements ─────────────────────────────────────────────────────────────
...
  /vessels/{vesselId}/dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get full dashboard data (cylinder status, alerts, spares)
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
  /vessels/{vesselId}/recalculate-rh:
    post:
      operationId: recalculateComponentRh
      tags: [dashboard]
      summary: Recalculate all component RH from movement log
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recalculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecalcResult"

  # ─── Valve Components ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    MonthlyRhEntry:
      type: object
      required: [id, vesselId, logDate, meTotalRh, monthlyRh, createdAt]
      properties:
        id:
          type: integer
        vesselId:
          type: integer
        logDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        meTotalRh:
          type: number
        monthlyRh:
          type: number
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    MonthlyRhInput:
      type: object
      required: [logDate, meTotalRh]
      properties:
        logDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        meTotalRh:
          type: number
          minimum: 0
        remarks:
          type: string

    # ─── Component ───────────────────────────────────────────────────────────
    Component:
      type: object
      required: [componentId, vesselId, componentType, condition, currentStatus, currentLocation, totalAccumulatedRh, createdAt]
      properties:
        componentId:
          type: string
        vesselId:
          type: integer
        componentType:
          type: string
        condition:
          type: string
        currentStatus:
          type: string
        currentLocation:
          type: string
        totalAccumulatedRh:
          type: number
        fittedAtMeRh:
          type: ["number", "null"]
        overhaulRh:
          type: ["integer", "null"]
          description: "Per-component overhaul RH limit (overrides vessel default when set)"
        warningRh:
          type: ["integer", "null"]
          description: "Per-component warning RH threshold (overrides vessel default when set)"
        effectiveOverhaulRh:
          type: integer
          description: "Resolved overhaul RH: individual → type-level → vessel default"
        effectiveWarningRh:
          type: integer
          description: "Resolved warning RH: individual → type-level → vessel default"
        remarks:
          type: ["string", "null"]
        liveRh:
          type: number
          description: "Live total RH = stored + (current_me - fitted_at) if in service"
        alertStatus:
          type: string
        createdAt:
          type: string
...
    Cylinder:
      type: object
      required: [id, vesselId, cylinderNumber]
      properties:
        id:
          type: integer
        vesselId:
          type: integer
        cylinderNumber:
          type: integer
        fittedComponentId:
          type: ["string", "null"]
        fittedAtMeRh:
          type: ["number", "null"]
        lastOverhaulRh:
          type: ["number", "null"]
        lastDismantlingRh:
          type: ["number", "null"]
        updatedAt:
          type: string
    CylinderUpdate:
      type: object
      properties:
        fittedComponentId:
          type: ["string", "null"]
        fittedAtMeRh:
          type: ["number", "null"]
        lastOverhaulRh:
          type: ["number", "null"]
        lastDismantlingRh:
          type: ["number", "null"]

    # ─── Movement ────────────────────────────────────────────────────────────
...
    CylinderComponentStatus:
      type: object
      required: [componentId, componentType, condition, totalRh, limit, alertStatus]
      properties:
        componentId:
          type: string
        componentType:
          type: string
        condition:
          type: string
        totalRh:
          type: number
          description: Total lifetime running hours of the component
        limit:
          type: number
          description: Overhaul threshold in hours for this component type
        alertStatus:
          type: string
    CylinderStatus:
      type: object
      required: [cylinder, components, rhSinceOverhaul, rhSinceDismantling, overallAlertStatus, dismantlingAlert]
      properties:
        cylinder:
          type: integer
        components:
          type: array
          items:
            $ref: "#/components/schemas/CylinderComponentStatus"
        rhSinceOverhaul:
          type: number
        rhSinceDismantling:
          type: number
        overallAlertStatus:
          type: string
        dismantlingAlert:
          type: string
    AlertItem:
      type: object
      required: [cylinder, componentId, type, status, totalRh, limit]
      properties:
        cylinder:
          type: integer
        componentId:
          type: string
        type:
          type: string
        status:
          type: string
        totalRh:
          type: number
        limit:
          type: number
        isChild:
          type: boolean
          description: "True if this alert is for a child component (nozzle, spring, etc.) rather than the main Unit/Fuel/Exhaust Valve component"
    DashboardData:
      type: object
      required: [currentMeRh, cylinderStatus, spareComponents, ashoreComponents, alerts, alertConfig]
      properties:
        currentMeRh:
          type: number
        cylinderStatus:
          type: array
          items:
            $ref: "#/components/schemas/CylinderStatus"
        spareComponents:
          type: array
          items:
            $ref: "#/components/schemas/Component"
        ashoreComponents:
          type: array
          items:
            $ref: "#/components/schemas/Component"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertItem"
        alertConfig:
          type: object
          required: [crownOverhaulRh, crownWarningRh, dismantlingWarningRh]
          properties:
            crownOverhaulRh:
              type: integer
            crownWarningRh:
              type: integer
            dismantlingWarningRh:
              type: integer
    RecalcResult:
      type: object
      required: [count, details]
      properties:
        count:
          type: integer
        details:
          type: array
          items:
            type: object
            required: [componentId, oldRh, newRh]
            properties:
              componentId:
                type: string
              oldRh:
                type: number
              newRh:
                type: number

    # ─── Valve Component ─────────────────────────────────────────────────────
...